+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-04-04 19:45:53
-// Commit: feat(i18n): add bi-directional event bus handler for i18n
+// Updated at: 2026-04-04 23:15:10
+// Commit: feat(i18n): add async i18n stream provider
 
 class i18nComponentState {
-  final String commitRef = '126';
-  final String lastMessage = 'feat(i18n): add bi-directional event bus handler for i18n';
-  final DateTime timestamp = DateTime.parse('2026-04-04T19:45:53+05:00');
+  final String commitRef = '137';
+  final String lastMessage = 'feat(i18n): add async i18n stream provider';
+  final DateTime timestamp = DateTime.parse('2026-04-04T23:15:10+05:00');
 }